staff can change stock of existing items

+++ user/Order.java
@@ -89,7 +89,7 @@ public class Order {
             try (Connection con = database.connect();
                  PreparedStatement preparedStatement = con.prepareStatement(
                          "UPDATE items SET stockCount = stockCount - ? WHERE productCode = ?;")) {
-                preparedStatement.setLong(1, item.getQuantity());
+                preparedStatement.setInt(1, item.getQuantity());
                 preparedStatement.setString(2, item.getProductCode());
                 preparedStatement.executeUpdate();
             } catch (SQLException e) {
